backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
numpy
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
flask-cors
flask-seeder
marshmallow
PyJWT